Apple Pay Gateway is missed during the wc rest api call on mollie return
The Apple pay gateway is removed from the payment methods list when
the device does not support it, it changed the implementation in fbab2444
then we added the session check at 3ec6edb1.

Those changes create the problem which the gateway is removed even during
the `wc-api` calls.

I introduced a more strict check to allow filtering of the gateways list
only if the current action is `woocommerce_payment_gateways` and it's
not a wc rest api call, not an ajax call. Those conditions includes
all of the previous ones.